@@ -3,7 +3,7 @@ package com.xattacker.android
 import android.graphics.Color
 import android.graphics.PointF
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.view.View
 import com.xattacker.android.view.polygon.*
 
